pyproject.toml
[build-system]
build-backend = "setuptools.build_meta"
requires = ["setuptools"]

[project]
name = "zoho-sdk"
authors = [
    {name = "Robert Huselius", email = "robert@huseli.us"},
]
description = ""
license = {file = "LICENSE"}
requires-python = ">=3.10"
dependencies = [
    "klaatu_python @ git+https://github.com/Eboreg/klaatu-python.git",
    "python-dotenv",
    "requests",
]
dynamic = ["version"]

[project.optional-dependencies]
dev = [
    "flake8",
    "ipdb",
    "ipython",
    "isort",
    "pylint",
    "types-requests",
]

[project.scripts]
zoho-get-token = "zoho.oauth2:get_oauth2_token_interactive"

[tool.setuptools.dynamic]
version = {attr = "__version__"}

[tool.pylint.main]
disable = [
    "broad-exception-caught",
    "import-outside-toplevel",
    "invalid-name",
    "invalid-str-returned",
    "missing-class-docstring",
    "missing-function-docstring",
    "missing-module-docstring",
    "too-few-public-methods",
    "too-many-ancestors",
    "too-many-arguments",
    "too-many-branches",
    "too-many-instance-attributes",
    "too-many-locals",
    "too-many-nested-blocks",
    "too-many-return-statements",
    "too-many-statements",
]
ignore = [".venv", "node_modules"]
ignore-patterns = ["^\\.#", ".*\\.pyi", "apps\\.py"]
max-line-length = 119

[tool.isort]
combine_as_imports = true
default_section = "THIRDPARTY"
include_trailing_comma = true
line_length = 79
multi_line_output = 3
lines_after_imports = 2
sections = [
    "FUTURE",
    "STDLIB",
    "THIRDPARTY",
    "FIRSTPARTY",
    "LOCALFOLDER",
]
skip = [".venv", "node_modules"]

[tool.mypy]
follow_imports = "skip"
show_error_codes = true
ignore_missing_imports = true
files = "."
check_untyped_defs = true
